# public/openapi/version-0.0-trv.yaml
openapi: 3.0.1
info:
  title: WIEDPUG API
  description: API Documentation for the WIEDPUG Network
  termsOfService: https://wiedpug.github.io/wiedpug
  contact:
    name: WIEDPUG Team
    url: https://wiedpug.github.io/wiedpug
    email: changethis@email.com
  license:
    name: Applicable License
    url: https://wiedpug.github.io/wiedpug
  version: v0.0
paths:
  /v0/test-resquests-verification:
    post:
      tags:
        - Test Requests Verification
      summary: Test Request Verification
      description: Sent by an organisation that has already requested a test by paper to the Test House, as a verification of the paper request
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestRequestsVerification'
            example:
              transmissionHeader:
                dateFormatLastRevised: '2020-07-15'
                dateTransmissionFileCreated: '2016-01-17'
                countryOfOrigin: AU
                documentOriginator:
                  organisationCode: AWT
                currentReceiver:
                  organisationCode: BCA
                versionNumber: 31
                softwareIdentifier:
                  softwareCompany: 8
                  version: 1082
                  programName: NETR
                abn: '72001967184'
              testRequestsVerificationDetails:
                - testDataHeader:
                    client:
                      organisationCode: AWT
                    lotIdentity: '567545'
                    weightNote:
                      prefix: L
                      suffix: '1234567'
                    bales: 1
                    gross:
                      value: 1
                      unit: 0
                    tare:
                      value: 1
                      unit: 0
                    weightUnit: 0
                    lotType: 3
                    centreAuction:
                      country: AU
                      centreCode: A
                    sourceCode: 1
                    isCompanyABNApplicable: false
                  testRequestVerification:
                    laboratory: 1
                    growerCode: G123
                    growerName: John Doe
                    coreSampleWeight:
                      value: 100
                      unit: 0
                    coreTestSamplingDate: '2024-04-15'
                    coresPerBale: 5
                    grabsPerBale: 10
                    isPreviousDataToBeUpdated: true
                    hashTotal: 123456
                    isYieldAndMicronTestRequired: true
                    isLengthAndStrengthTestRequired: false
                    isLengthOnlyTestRequired: true
                    isColourTestRequired: false
                    isLaserscanTestRequired: true
                    isAirflowTestRequired: false
                    isOFDATestRequired: true
                    isDarkAndMedullatedFibreTestRequired: false
                    isPesticideResidueTestRequired: false
                  contractLimitRequirements:
                    micronMin: 15.5
                    micronMax: 25
                    colourYZMin: 1.2
                    colourYZMax: 2.1
                    vmbMin: 0.85
                    vmbMax: 1.25
                    woolBaseMin: 50
                    woolBaseMax: 65
                    dichloromethaneDryBasisMin: 0.05
                    dichloromethaneDryBasisMax: 0.15
                    conditioningMassMin:
                      value: 500
                      unit: 0
                    conditioningMassMax:
                      value: 1500
                      unit: 0
                    lacBarbeMin: 10
                    lacBarbeMax: 30
                    lacHauteurMin: 15
                    lacHauteurMax: 45
                    lacHauteurUnder20Min: 5
                    lacHauteurUnder20Max: 15
                    lacHauteurUnder40Min: 8
                    lacHauteurUnder40Max: 25
                    lacHauteurCVPercentMin: 10
                    lacHauteurCVPercentMax: 20
                  darkAndMedullatedFibreRiskDetails:
                    declaredBreed: A
                    declaredWoolSubCategory: 'N'
                    declaredWoolCategory: P
                    isContactingWithExoticBreeds: true
                    isCrutched: false
                    isCrutchedWithin3MonthsOfShearing: true
                    sex: R
                    ageFrom: 2
                    ageTo: 5
                - testDataHeader:
                    client:
                      organisationCode: AWT
                    lotIdentity: '567545'
                    weightNote:
                      prefix: L
                      suffix: '1234567'
                    bales: 1
                    gross:
                      value: 1
                      unit: 0
                    tare:
                      value: 1
                      unit: 0
                    weightUnit: 0
                    lotType: 3
                    centreAuction:
                      country: AU
                      centreCode: BA
                    sourceCode: 2
                    isCompanyABNApplicable: false
                  testRequestVerification:
                    laboratory: 1
                    growerCode: G4643
                    growerName: Alice Stonewood
                    coreSampleWeight:
                      value: 100
                      unit: 0
                    coreTestSamplingDate: '2024-04-25'
                    coresPerBale: 4
                    grabsPerBale: 20
                    isPreviousDataToBeUpdated: true
                    hashTotal: 125534
                    isYieldAndMicronTestRequired: true
                    isLengthAndStrengthTestRequired: false
                    isLengthOnlyTestRequired: false
                    isColourTestRequired: false
                    isLaserscanTestRequired: false
                    isAirflowTestRequired: false
                    isOFDATestRequired: false
                    isDarkAndMedullatedFibreTestRequired: true
                    isPesticideResidueTestRequired: true
                  contractLimitRequirements:
                    micronMin: 12.3
                    micronMax: 18.6
                    colourYZMin: 1.6
                    colourYZMax: 2.5
                    vmbMin: 0.66
                    vmbMax: 1.8
                    woolBaseMin: 40
                    woolBaseMax: 55
                    dichloromethaneDryBasisMin: 0.06
                    dichloromethaneDryBasisMax: 0.12
                    conditioningMassMin:
                      value: 260
                      unit: 0
                    conditioningMassMax:
                      value: 1200
                      unit: 0
                    lacBarbeMin: 17
                    lacBarbeMax: 56
                    lacHauteurMin: 12
                    lacHauteurMax: 36
                    lacHauteurUnder20Min: 3
                    lacHauteurUnder20Max: 12
                    lacHauteurUnder40Min: 9
                    lacHauteurUnder40Max: 15
                    lacHauteurCVPercentMin: 13
                    lacHauteurCVPercentMax: 27
                  darkAndMedullatedFibreRiskDetails:
                    declaredBreed: A
                    declaredWoolSubCategory: 'N'
                    declaredWoolCategory: B
                    isContactingWithExoticBreeds: false
                    isCrutched: false
                    isCrutchedWithin3MonthsOfShearing: true
                    sex: E
                    ageFrom: 1
                    ageTo: 4
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResult'
              example: {}
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.1
                title: One or more validation errors occurred
                status: 400
                detail: ''
                errors:
                  errors:
                    Requests[0].TransmissionHeader.DateFormatLastRevised:
                      - The field DateFormatLastRevised is required
                      - The field DateFormatLastRevised must match regulart expression pattern...
                    Requests[0].TransmissionHeader.CountryOfOrigin:
                      - The field CountryOfOrigin is required
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.2
                title: Unauthorized
                status: 401
                detail: Missing authentication credentials
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.4
                title: You don't have permission to access this resource
                status: 403
                detail: You don't have a 'read' access to access this resource
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#name-500-internal-server-error
                title: An error occurred while processing your required
                status: 500
                detail: Error details..
  /v0/test-resquests-verification/data:
    post:
      tags:
        - Test Requests Verification
      summary: Retrieves the test requests verification
      operationId: RequestForTestRequestsVerification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestForData'
            example:
              saleIdentity:
                sellingCentreType: AU_M
                saleNumber: 18
              organisationRequested:
                organisationCode: ANF
              saleDateRequested: '2023-10-31'
              isUsingNetworkDateTime: false
              requestType: 4
              woolTypeGroup: MFLC
              season: 23
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRequestsVerificationApiResult'
              example:
                data:
                  transmissionHeader:
                    dateFormatLastRevised: '2020-07-15'
                    dateTransmissionFileCreated: '2016-01-17'
                    countryOfOrigin: AU
                    documentOriginator:
                      organisationCode: AWT
                    currentReceiver:
                      organisationCode: BCA
                    versionNumber: 31
                    softwareIdentifier:
                      softwareCompany: 8
                      version: 1082
                      programName: NETR
                    abn: '72001967184'
                  testRequestsVerificationDetails:
                    - testDataHeader:
                        client:
                          organisationCode: AWT
                        lotIdentity: '567545'
                        weightNote:
                          prefix: L
                          suffix: '1234567'
                        bales: 1
                        gross:
                          value: 1
                          unit: 0
                        tare:
                          value: 1
                          unit: 0
                        weightUnit: 0
                        lotType: 3
                        centreAuction:
                          country: AU
                          centreCode: A
                        sourceCode: 1
                        isCompanyABNApplicable: false
                      testRequestVerification:
                        laboratory: 1
                        growerCode: G123
                        growerName: John Doe
                        coreSampleWeight:
                          value: 20
                          unit: 0
                        coreTestSamplingDate: '2024-04-15'
                        coresPerBale: 5
                        grabsPerBale: 10
                        isPreviousDataToBeUpdated: true
                        hashTotal: 123456
                        isYieldAndMicronTestRequired: true
                        isLengthAndStrengthTestRequired: false
                        isLengthOnlyTestRequired: true
                        isColourTestRequired: false
                        isLaserscanTestRequired: true
                        isAirflowTestRequired: false
                        isOFDATestRequired: true
                        isDarkAndMedullatedFibreTestRequired: false
                        isPesticideResidueTestRequired: false
                      contractLimitRequirements:
                        micronMin: 15.5
                        micronMax: 25
                        colourYZMin: 1.2
                        colourYZMax: 2.1
                        vmbMin: 0.85
                        vmbMax: 1.25
                        woolBaseMin: 50
                        woolBaseMax: 65
                        dichloromethaneDryBasisMin: 0.05
                        dichloromethaneDryBasisMax: 0.15
                        conditioningMassMin:
                          value: 500
                          unit: 0
                        conditioningMassMax:
                          value: 1500
                          unit: 0
                        lacBarbeMin: 10
                        lacBarbeMax: 30
                        lacHauteurMin: 15
                        lacHauteurMax: 45
                        lacHauteurUnder20Min: 5
                        lacHauteurUnder20Max: 15
                        lacHauteurUnder40Min: 8
                        lacHauteurUnder40Max: 25
                        lacHauteurCVPercentMin: 10
                        lacHauteurCVPercentMax: 20
                      darkAndMedullatedFibreRiskDetails:
                        declaredBreed: A
                        declaredWoolSubCategory: 'N'
                        declaredWoolCategory: P
                        isContactingWithExoticBreeds: true
                        isCrutched: false
                        isCrutchedWithin3MonthsOfShearing: true
                        sex: R
                        ageFrom: 2
                        ageTo: 5
                    - testDataHeader:
                        client:
                          organisationCode: AWT
                        lotIdentity: '567545'
                        weightNote:
                          prefix: L
                          suffix: '1234567'
                        bales: 1
                        gross:
                          value: 1
                          unit: 0
                        tare:
                          value: 1
                          unit: 0
                        weightUnit: 0
                        lotType: 3
                        centreAuction:
                          country: AU
                          centreCode: A
                        sourceCode: 2
                        isCompanyABNApplicable: false
                      testRequestVerification:
                        laboratory: 1
                        growerCode: G4643
                        growerName: Alice Stonewood
                        coreSampleWeight:
                          value: 164
                          unit: 0
                        coreTestSamplingDate: '2024-04-25'
                        coresPerBale: 4
                        grabsPerBale: 20
                        isPreviousDataToBeUpdated: true
                        hashTotal: 125534
                        isYieldAndMicronTestRequired: true
                        isLengthAndStrengthTestRequired: false
                        isLengthOnlyTestRequired: false
                        isColourTestRequired: false
                        isLaserscanTestRequired: false
                        isAirflowTestRequired: false
                        isOFDATestRequired: false
                        isDarkAndMedullatedFibreTestRequired: true
                        isPesticideResidueTestRequired: true
                      contractLimitRequirements:
                        micronMin: 12.3
                        micronMax: 18.6
                        colourYZMin: 1.6
                        colourYZMax: 2.5
                        vmbMin: 0.66
                        vmbMax: 1.8
                        woolBaseMin: 40
                        woolBaseMax: 55
                        dichloromethaneDryBasisMin: 0.06
                        dichloromethaneDryBasisMax: 0.12
                        conditioningMassMin:
                          value: 126
                          unit: 0
                        conditioningMassMax:
                          value: 1200
                          unit: 0
                        lacBarbeMin: 17
                        lacBarbeMax: 56
                        lacHauteurMin: 12
                        lacHauteurMax: 36
                        lacHauteurUnder20Min: 3
                        lacHauteurUnder20Max: 12
                        lacHauteurUnder40Min: 9
                        lacHauteurUnder40Max: 15
                        lacHauteurCVPercentMin: 13
                        lacHauteurCVPercentMax: 27
                      darkAndMedullatedFibreRiskDetails:
                        declaredBreed: A
                        declaredWoolSubCategory: 'N'
                        declaredWoolCategory: B
                        isContactingWithExoticBreeds: false
                        isCrutched: false
                        isCrutchedWithin3MonthsOfShearing: true
                        sex: E
                        ageFrom: 1
                        ageTo: 4
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.1
                title: One or more validation errors occurred
                status: 400
                detail: ''
                errors:
                  errors:
                    Requests[0].TransmissionHeader.DateFormatLastRevised:
                      - The field DateFormatLastRevised is required
                      - The field DateFormatLastRevised must match regulart expression pattern...
                    Requests[0].TransmissionHeader.CountryOfOrigin:
                      - The field CountryOfOrigin is required
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.2
                title: Unauthorized
                status: 401
                detail: Missing authentication credentials
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.4
                title: You don't have permission to access this resource
                status: 403
                detail: You don't have a 'read' access to access this resource
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#section-15.5.4
                title: Resource not found
                status: 404
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiErrorResult'
              example:
                type: https://datatracker.ietf.org/doc/html/rfc9110#name-500-internal-server-error
                title: An error occurred while processing your required
                status: 500
                detail: Error details..
components:
  schemas:
    ApiErrorResult:
      required:
        - status
        - title
      type: object
      properties:
        type:
          type: string
          description: "A URI reference [RFC3986] that identifies the problem type.\r\nThis specification encourages that, when dereferenced, it provide human-readable documentation for the problem type (e.g., using HTML [W3C.REC-html5-20141028]). \r\nWhen this member is not present, its value is assumed to be \"about:blank\".      \r\ne.g. For 400 Bad Request, the \"type\" value can be \"https://tools.ietf.org/html/rfc9110#section-15.5.1\""
          nullable: true
        title:
          minLength: 1
          type: string
          description: "A short, human-readable summary of the problem type. \r\nIt SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization(e.g., using proactive content negotiation; see[RFC7231], Section 3.4).\r\ne.g. \"One or more validation errors occurred.\""
        status:
          type: integer
          description: The HTTP status code ([RFC7231], Section 6) generated by the origin server for this occurrence of the problem.
          format: int32
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem. e.g. 400, 403, or business relared problem etc.
          nullable: true
        errors:
          $ref: '#/components/schemas/CustomErrors'
      additionalProperties: false
      description: A machine-readable format for specifying errors in HTTP API responses based on https://tools.ietf.org/html/rfc7807.
    ApiMessage:
      type: object
      properties:
        message:
          type: string
          nullable: true
        warningCodes:
          type: array
          items:
            $ref: '#/components/schemas/WarningCode'
          nullable: true
        informationCodes:
          type: array
          items:
            $ref: '#/components/schemas/InformationCode'
          nullable: true
      additionalProperties: false
    ApiResult:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/ApiMessage'
        data:
          description: Contains response data. e.g. Auction Catalogue
          nullable: true
      additionalProperties: false
      description: A standard API response object used for all successful API responses
...
    CustomErrors:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
            nullable: true
          description: 'Dictionary of errors in <PropertyName, ErrorDetails[]> format. e.g. errors: { "name": ["Product name is required"], "category": ["Category name cannot be longer than 20 characters."]  }'
          nullable: true
        errorCodes:
          type: array
          items:
            $ref: '#/components/schemas/ErrorCode'
          nullable: true
      additionalProperties: false
...
    Organisation:
      required:
        - organisationCode
      type: object
      properties:
        organisationCode:
          maxLength: 3
          minLength: 1
          type: string
        centre:
          $ref: '#/components/schemas/Centre'
      additionalProperties: false
    RequestForData:
      required:
        - requestType
      type: object
      properties:
        saleIdentity:
          $ref: '#/components/schemas/SaleIdentity'
        organisationRequested:
          $ref: '#/components/schemas/Organisation'
        saleDateRequested:
          maxLength: 10
          minLength: 1
          pattern: /(?:[1-9]\d{3}-(?:(?:0[1-9]|1[0-2])-(?:0[1-9]|1\d|2[0-8])|(?:0[13-9]|1[0-2])-(?:29|30)|(?:0[13578]|1[02])-31)|(?:[1-9]\d(?:0[48]|[2468][048]|[13579][26])|(?:[2468][048]|[13579][26])00)-02-29)$/
          type: string
          description: Date value in ISO 8601 standard. e.g. 2024-03-21
          format: date
          nullable: true
        startDateTimeStatusRequest:
          maxLength: 20
          minLength: 1
          pattern: ^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])T([01]\d|2[0-3]):([0-5]\d):([0-5]\d)Z$
          type: string
          description: Date and Time value in ISO 8601 standard UTC format. e.g. 2024-03-21T19:25:04Z
          format: date-time
          nullable: true
        endDateTimeStatusRequest:
          maxLength: 20
          minLength: 1
          pattern: ^\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])T([01]\d|2[0-3]):([0-5]\d):([0-5]\d)Z$
          type: string
          description: Date and Time with offset value in ISO 8601 standard UTC format. e.g. 2024-03-21T19:25:04Z
          format: date-time
          nullable: true
        isUsingNetworkDateTime:
          maxLength: 5
          minLength: 4
          type: boolean
          description: "A boolean value to indicate whethear to use network datetime.\r\ntrue: use network datetime.\r\nfalse: not use."
          nullable: true
        requestType:
          $ref: '#/components/schemas/RequestTypeForData'
        woolTypeGroup:
          $ref: '#/components/schemas/WoolTypeGroup'
        season:
          pattern: '-?\d{1,2}'
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    RequestTypeForData:
      enum:
        - 1
        - 3
        - 4
      type: integer
      description: '1: Private Data Request<br />3: Data Re-Request<br />4: Broadcast Data Request'
      format: int32
    SaleIdentity:
      required:
        - saleNumber
        - sellingCentreType
      type: object
      properties:
        sellingCentreType:
          $ref: '#/components/schemas/SellingCentreType'
        saleNumber:
          maxLength: 2
          minLength: 1
          pattern: '-?\d{1,2}'
          type: integer
          format: int32
      additionalProperties: false
...
    TestRequestsVerification:
      required:
        - transmissionHeader
      type: object
      properties:
        transmissionHeader:
          $ref: '#/components/schemas/TransmissionHeader'
        transferConfig:
          $ref: '#/components/schemas/TransferConfig'
        testRequestsVerificationDetails:
          type: array
          items:
            $ref: '#/components/schemas/TestRequestsVerificationDetail'
          nullable: true
      additionalProperties: false
    TestRequestsVerificationApiResult:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/ApiMessage'
        data:
          $ref: '#/components/schemas/TestRequestsVerification'
      additionalProperties: false
      description: A standard API response object used for all successful API responses
    TestRequestsVerificationDetail:
      required:
        - testDataHeader
        - testRequestVerification
      type: object
      properties:
        testDataHeader:
          $ref: '#/components/schemas/TestDataHeader'
        testRequestVerification:
          $ref: '#/components/schemas/TestRequestVerification'
        contractLimitRequirements:
          $ref: '#/components/schemas/ContractLimitRequirements'
        darkAndMedullatedFibreRiskDetails:
          $ref: '#/components/schemas/TestRequestVerificationDarkAndMedullatedFibreRiskDetails'
        bales:
          type: array
          items:
            $ref: '#/components/schemas/Bale'
          nullable: true
      additionalProperties: false
    TransferConfig:
      required:
        - transferType
      type: object
      properties:
        transferType:
          $ref: '#/components/schemas/TransferType'
        finalReceiverList:
          type: array
          items:
            $ref: '#/components/schemas/Organisation'
          nullable: true
      additionalProperties: false
      description: "Used to identify document transfer type: Restrict, Unrestric, and Exlcude.\r\nOnly used for data retrieval. Do not include it in the response."
    TransferType:
      enum:
        - Unrestrict
        - Restrict
        - Exclude
      type: string
      description: 'Unrestrict: Send document to all members<br />Restrict: Send document to appointed members<br />Exclude: Send to all members excluding the appointed members'
    TransmissionHeader:
      required:
        - abn
        - countryOfOrigin
        - currentReceiver
        - dateFormatLastRevised
        - dateTransmissionFileCreated
        - documentOriginator
      type: object
      properties:
        dateFormatLastRevised:
          maxLength: 10
          minLength: 1
          pattern: /(?:[1-9]\d{3}-(?:(?:0[1-9]|1[0-2])-(?:0[1-9]|1\d|2[0-8])|(?:0[13-9]|1[0-2])-(?:29|30)|(?:0[13578]|1[02])-31)|(?:[1-9]\d(?:0[48]|[2468][048]|[13579][26])|(?:[2468][048]|[13579][26])00)-02-29)$/
          type: string
          description: Date value in ISO standard UTC date format. e.g. 2024-03-21 (YYYY-MM-DD)
          format: date
        dateTransmissionFileCreated:
          maxLength: 10
          minLength: 1
          pattern: /(?:[1-9]\d{3}-(?:(?:0[1-9]|1[0-2])-(?:0[1-9]|1\d|2[0-8])|(?:0[13-9]|1[0-2])-(?:29|30)|(?:0[13578]|1[02])-31)|(?:[1-9]\d(?:0[48]|[2468][048]|[13579][26])|(?:[2468][048]|[13579][26])00)-02-29)$/
          type: string
          description: Date value in ISO standard UTC date format. e.g. 2024-03-21 (YYYY-MM-DD)
          format: date
        countryOfOrigin:
          $ref: '#/components/schemas/CountryCode'
        documentOriginator:
          $ref: '#/components/schemas/Organisation'
        currentReceiver:
          $ref: '#/components/schemas/Organisation'
        versionNumber:
          pattern: '-?\d{1,2}'
          type: integer
          description: Two digits. e.g. 99
          format: int32
          nullable: true
        softwareIdentifier:
          $ref: '#/components/schemas/SoftwareIdentifier'
        abn:
          maxLength: 11
          minLength: 1
          pattern: ^[0-9]+$
          type: string
          description: Australian Business Number, 11 digits string
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: 'JWT Access Token that is added in the Authorization header of the HTTP request. Example: "Authorization: Bearer {token}". The JWT serves as a credential that proves the identity of the client to the server.'
      name: Authorization
      in: header
    oauth2:
      type: oauth2
      description: Authentication for client applications by providing Client ID and Client Secret for an JWT Access Token.
      flows:
        clientCredentials:
          tokenUrl: /auth-server/oauth/token
          scopes: {}
security:
  - Bearer: []